moved repos that all attack and defend into separate folders and namespaces
