+++ WebApp/Startup.cs
@@ -9,7 +9,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using DataLayer.Interfaces;
-using DataLayer.Repositories;
 
 namespace WebApp
 {
@@ -25,8 +24,8 @@ namespace WebApp
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddScoped<ITitleRepository, TitleRepository>();
-            services.AddScoped<ICustomerRepository, CustomerRepositoryEF>();
+            services.AddScoped<ITitleRepository, DataLayer.Repositories.Attack.TitleRepository>();
+            services.AddScoped<ICustomerRepository, DataLayer.Repositories.DefendWithEF.CustomerRepository>();
 
             services.AddControllersWithViews();
         }
